@@ -21,137 +21,67 @@
     "postbuild": "workbox injectManifest ./config/workbox-config.js",
     "compress": "bread-compressor -a gzip www",
     "build:prod": "npm run build:aot && node ./cache-busting.js && npm run postbuild && npm run compress",
+    "build:prod2": "ionic-app-scripts build --prod && node ./cache-busting.js && npm run postbuild && npm run compress",
     "deploy": "firebase deploy --project ma-cave-15a66"
   },
   "dependencies": {
-    "@angular/cli": "6.0.5",
-    "@angular/common": "6.0.3",
-    "@angular/compiler": "6.0.3",
-    "@angular/compiler-cli": "6.0.3",
-    "@angular/core": "6.0.3",
-    "@angular/forms": "6.0.3",
-    "@angular/http": "6.0.3",
-    "@angular/platform-browser": "6.0.3",
-    "@angular/platform-browser-dynamic": "6.0.3",
-    "@angular/platform-server": "6.0.3",
-    "@angular/tsc-wrapped": "4.4.6",
-    "@ionic-native/android-permissions": "4.7.0",
-    "@ionic-native/barcode-scanner": "4.7.0",
-    "@ionic-native/camera": "4.7.0",
-    "@ionic-native/core": "4.7.0",
-    "@ionic-native/device-feedback": "4.7.0",
-    "@ionic-native/facebook": "4.7.0",
-    "@ionic-native/file": "4.7.0",
-    "@ionic-native/file-chooser": "4.7.0",
-    "@ionic-native/file-path": "4.7.0",
-    "@ionic-native/google-plus": "4.7.0",
-    "@ionic-native/native-storage": "4.7.0",
-    "@ionic-native/network": "4.7.0",
-    "@ionic-native/splash-screen": "4.7.0",
-    "@ionic-native/status-bar": "4.7.0",
-    "@ionic-native/transfer": "3.14.0",
+    "@angular/cli": "6.1.0-rc.3",
+    "@angular/common": "6.1.0-rc.3",
+    "@angular/compiler": "6.1.0-rc.3",
+    "@angular/compiler-cli": "6.1.0-rc.3",
+    "@angular/core": "6.1.0-rc.3",
+    "@angular/forms": "6.1.0-rc.3",
+    "@angular/http": "6.1.0-rc.3",
+    "@angular/platform-browser": "6.1.0-rc.3",
+    "@angular/platform-browser-dynamic": "6.1.0-rc.3",
+    "@angular/platform-server": "6.1.0-rc.3",
+    "@angular/tsc-wrapped": "4.4.7",
     "@ionic/storage": "2.1.3",
     "@ngrx/effects": "6.0.1",
     "@ngrx/store": "6.0.1",
     "@ngrx/store-devtools": "6.0.1",
     "@ngx-translate/core": "10.0.2",
     "@ngx-translate/http-loader": "3.0.1",
-    "angular2-uuid": "1.1.1",
-    "angularfire2": "5.0.0-rc.10",
-    "bootstrap": "4.1.1",
+    "angularfire2": "5.0.0-rc.11",
+    "bootstrap": "4.1.2",
     "chart.js": "2.7.2",
-    "cordova-android": "7.1.0",
-    "cordova-plugin-android-permissions": "1.0.0",
-    "cordova-plugin-device": "2.0.2",
-    "cordova-plugin-facebook4.4": "1.0.1",
-    "cordova-plugin-googleplus": "5.3.0",
-    "cordova-plugin-image-picker": "1.1.3",
-    "cordova-plugin-network-information": "2.0.1",
-    "cordova-plugin-splashscreen": "5.0.2",
-    "cordova-plugin-statusbar": "2.4.2",
-    "cordova-plugin-velda-devicefeedback": "0.0.2",
-    "cordova-plugin-whitelist": "1.3.3",
-    "firebase": "5.0.4",
+    "firebase": "5.2.0",
     "firebaseui": "3.1.1",
     "firebaseui-angular": "3.1.1",
     "ionic-angular": "3.9.2",
-    "ionic-plugin-keyboard": "2.2.1",
     "lodash": "4.17.10",
     "loglevel": "1.6.1",
     "ng2-charts": "1.6.0",
-    "phonegap-plugin-barcodescanner": "7.1.2",
-    "rxjs": "6.2.0",
-    "rxjs-compat": "6.2.0",
-    "rxjs-tslint": "0.1.4",
+    "rxjs": "6.2.2",
+    "rxjs-compat": "6.2.2",
+    "rxjs-tslint": "0.1.5",
     "text-encoding": "0.6.4",
-    "workbox-sw": "3.2.0",
+    "workbox-sw": "3.4.1",
     "zone.js": "0.8.26"
   },
   "devDependencies": {
-    "@ionic/app-scripts": "3.1.9",
-    "@types/jasmine": "2.8.7",
-    "@types/karma": "1.7.3",
-    "@types/node": "10.1.3",
-    "bread-compressor-cli": "1.0.0",
+    "@ionic/app-scripts": "3.1.11",
+    "@types/bluebird": "3.5.21",
+    "@types/jasmine": "2.8.8",
+    "@types/karma": "1.7.5",
+    "@types/node": "10.5.2",
+    "bread-compressor-cli": "1.0.4",
     "http-server": "0.11.1",
     "ionic": "3.20.0",
     "jasmine-core": "3.1.0",
     "jasmine-spec-reporter": "4.2.1",
-    "ngrx-store-freeze": "0.2.3",
+    "ngrx-store-freeze": "0.2.4",
     "promise-polyfill": "8.0.0",
     "purify-css": "1.2.5",
     "rev-hash": "2.0.0",
-    "typescript": ">=2.7.2 <2.8",
+    "typescript": ">=2.7.2 and <2.8.0",
     "webpack-manifest-plugin": "2.0.3",
-    "workbox-cli": "3.2.0"
+    "workbox-cli": "3.4.1"
   },
-  "cordovaPlugins": [
-    "cordova-plugin-device",
-    "cordova-plugin-image-picker",
-    "cordova-plugin-splashscreen",
-    "cordova-plugin-statusbar",
-    "cordova-plugin-whitelist",
-    "ionic-plugin-keyboard"
-  ],
   "config": {
     "ionic_copy": "./config/copy.config.js",
     "ionic_sass": "./config/sass.config.js",
     "ionic_webpack": "./config/webpack.config.js"
   },
-  "cordovaPlatforms": [
-    "ios",
-    {
-      "platform": "ios",
-      "version": "",
-      "locator": "ios"
-    }
-  ],
-  "description": "L'assistant caviste: ne perdez plus vos bouteilles !",
-  "cordova": {
-    "platforms": [
-      "android"
-    ],
-    "plugins": {
-      "cordova-plugin-device": {},
-      "cordova-plugin-splashscreen": {},
-      "cordova-plugin-statusbar": {},
-      "cordova-plugin-whitelist": {},
-      "ionic-plugin-keyboard": {},
-      "phonegap-plugin-barcodescanner": {
-        "CAMERA_USAGE_DESCRIPTION": " "
-      },
-      "cordova-plugin-velda-devicefeedback": {},
-      "cordova-plugin-android-permissions": {},
-      "cordova-plugin-network-information": {},
-      "cordova-plugin-image-picker": {},
-      "cordova-plugin-facebook4.4": {
-        "APP_ID": "253712085111353",
-        "APP_NAME": "cellar-explorer"
-      },
-      "com.megster.cordova.FileChooser": {},
-      "cordova-plugin-googleplus": {
-        "REVERSED_CLIENT_ID": "58435015061-8bnsnki77q4ffi25ph5plr6m694866vd.apps.googleusercontent.com"
-      }
-    }
-  }
+  "description": "L'assistant caviste: ne perdez plus vos bouteilles !"
 }
