@@ -3,11 +3,6 @@ import {IonicPageModule} from 'ionic-angular';
 import {FormsModule} from '@angular/forms';
 import {SharedModule} from '../../components/shared.module';
 import {AdminPage} from './admin-page';
-import {Transfer, TransferObject} from '@ionic-native/transfer';
-import {FilePath} from '@ionic-native/file-path';
-import {File} from '@ionic-native/file';
-import {BarcodeScanner} from '@ionic-native/barcode-scanner';
-import {FileChooser} from '@ionic-native/file-chooser';
 import {CommonModule} from '@angular/common';
 import {TranslateModule} from '@ngx-translate/core';
 import {ImportProvider} from '../../providers/import/import';
@@ -30,13 +25,7 @@ import {ImportProvider} from '../../providers/import/import';
               AdminPage
             ],
             providers: [
-              Transfer,
-              TransferObject,
-              FileChooser,
-              File,
-              FilePath,
-              ImportProvider,
-              BarcodeScanner
+              ImportProvider
             ]
           })
 export class AdminPageModule {
