@@ -1,7 +1,5 @@
 import {Component} from '@angular/core';
 import {IonicPage, Loading, LoadingController, NavController, Platform} from 'ionic-angular';
-import {FileChooser} from '@ionic-native/file-chooser';
-import {FilePath} from '@ionic-native/file-path';
 import {BottlePersistenceService} from '../../service/bottle-persistence.service';
 import {NotificationService} from '../../service/notification.service';
 import {ImportProvider} from '../../providers/import/import';
@@ -9,11 +7,10 @@ import {ApplicationState} from '../../app/state/app.state';
 import {Store} from '@ngrx/store';
 import {BottlesQuery} from '../../app/state/bottles.state';
 import {DeleteAccountAction, LogoutAction} from '../../app/state/shared.actions';
-import {SharedQuery, SharedState} from '../../app/state/shared.state';
+import {SharedQuery} from '../../app/state/shared.state';
 import {Observable} from 'rxjs';
 import {User} from '../../model/user';
 import {take} from 'rxjs/operators';
-import {HomePage} from '../../app/home/home';
 
 /**
  * Generated class for the UploadBottles page.
@@ -36,15 +33,13 @@ export class AdminPage {
   private user$: Observable<User>;
 
   constructor(private navCtrl: NavController,
-              private filepath: FilePath,
-              private fileChooser: FileChooser,
               private notificationService: NotificationService,
               private bottleService: BottlePersistenceService,
               private platform: Platform,
               private loadingController: LoadingController,
               private importProvider: ImportProvider,
               private store: Store<ApplicationState>) {
-    this.user$=this.store.select(SharedQuery.getLoginUser);
+    this.user$ = this.store.select(SharedQuery.getLoginUser);
   }
 
   deleteAccount() {
@@ -68,32 +63,8 @@ export class AdminPage {
     //  , 100);
   }
 
-  public platformIsCordova(): boolean {
-    return this.platform.is('cordova');
-  }
-
   public switchAdvancedOptions() {
     this.optionsVisibles = !this.optionsVisibles;
-  }
-
-  /**
-   * si Cordova seulement
-   **/
-  public chooseFile() {
-    this.fileChooser.open()
-      .then(uri => {
-        this.filepath.resolveNativePath(uri).then((nativepath) => {
-          (<any>window).resolveLocalFileSystemURL(nativepath, (res) => {
-            res.file((resFile) => {
-              this.setupUpload(resFile);
-            });
-          });
-        })
-          .catch(error => this.notificationService.error('uri incorrecte ' + uri + ' error:' + error));
-      })
-      .catch(error => {
-        this.notificationService.error('Erreur lors du choix du fichier', error);
-      });
   }
 
   public readBrowserFile(event: any) {
