@@ -1,9 +1,7 @@
 import {Component} from '@angular/core';
 import {Platform} from 'ionic-angular';
 import {TranslateService} from '@ngx-translate/core';
-import {SplashScreen} from '@ionic-native/splash-screen';
 import {HomePage} from './home/home';
-import {logInfo} from '../utils';
 
 @Component({
              templateUrl: 'app.html'
@@ -12,15 +10,8 @@ export class MyCaveApp {
   selectedTheme = 'ionic-theme';
   rootPage = HomePage;
 
-  constructor(platform: Platform, translate: TranslateService,
-              splashScreen: SplashScreen) {
+  constructor(platform: Platform, translate: TranslateService) {
     platform.ready().then(() => {
-      // Okay, so the platform is ready and our plugins are available.
-      // Here you can do any higher level native things you might need.
-      setTimeout(() => {
-        splashScreen.hide();
-      }, 100);
-
       // this language will be used as a fallback when a translation isn't found in the current language
       translate.setDefaultLang('fr');
 
