@@ -2,6 +2,7 @@ import {Action} from '@ngrx/store';
 import {UserPreferences} from '../../model/user-preferences';
 import {User} from '../../model/user';
 import {LOGINTYPE} from '../../service/login/login.service';
+import {BOTTLE_ITEM_TYPE} from './shared.state';
 
 export enum SharedActionTypes {
   LoadSharedActionType = '[shared] - load',
@@ -12,7 +13,7 @@ export enum SharedActionTypes {
   LoginActionSuccessType = '[shared] - login success',
   LogoutActionType = '[shared] - logout',
   DeleteAccountActionType = '[shared] - delete account',
-  UpdateThemeActionType = '[shared] - update theme',
+  UpdatePrefsActionType = '[shared] - update preferences',
   UpdateMostUsedQueriesActionType = '[shared] - update most used queries',
 }
 
@@ -24,7 +25,7 @@ export type SharedActions = LoadSharedAction
   | LoginFailAction
   | LogoutAction
   | DeleteAccountAction
-  | UpdateThemeAction
+  | UpdatePrefsAction
   | UpdateMostUsedQueriesAction;
 
 /**
@@ -48,10 +49,10 @@ export class LoadSharedFailedAction implements Action {
   }
 }
 
-export class UpdateThemeAction implements Action {
-  readonly type = SharedActionTypes.UpdateThemeActionType;
+export class UpdatePrefsAction implements Action {
+  readonly type = SharedActionTypes.UpdatePrefsActionType;
 
-  constructor(public theme: string) {
+  constructor(public theme: string, public itemType: BOTTLE_ITEM_TYPE) {
   }
 }
 
