@@ -4,16 +4,23 @@ import {User} from '../../model/user';
 import {SearchCriteria} from '../../model/search-criteria';
 import {SharedActions, SharedActionTypes} from './shared.actions';
 
+export enum BOTTLE_ITEM_TYPE {
+  STANDARD = 'standard',
+  LARGE = 'large'
+};
+
 export interface SharedState {
   mostUsedQueries: SearchCriteria[];
   theme: string;
+  bottleItemType: BOTTLE_ITEM_TYPE,
   user: User;
 }
 
 const INITIAL_STATE: SharedState = {
   mostUsedQueries: [],
   theme: 'default',
-  user: undefined
+  bottleItemType: BOTTLE_ITEM_TYPE.STANDARD,
+  user: undefined,
 };
 
 export namespace SharedQuery {
@@ -28,6 +35,7 @@ export function sharedStateReducer(state: SharedState = INITIAL_STATE, action: S
         return {
           ...state,
           mostUsedQueries: action.prefs.mostUsedQueries,
+          bottleItemType: action.prefs.itemType,
           theme: action.prefs.theme
         };
       } else {
@@ -36,14 +44,22 @@ export function sharedStateReducer(state: SharedState = INITIAL_STATE, action: S
     }
 
     case SharedActionTypes.LoginActionSuccessType: {
-        return {
-          ...state,
-          user: action.user
-        };
+      return {
+        ...state,
+        user: action.user
+      };
     }
 
     case SharedActionTypes.LogoutActionType: {
-        return INITIAL_STATE;
+      return INITIAL_STATE;
+    }
+
+    case SharedActionTypes.UpdatePrefsActionType: {
+      return {
+        ...state,
+        theme: action.theme,
+        bottleItemType: action.itemType
+      };
     }
 
     default:
