@@ -8,7 +8,8 @@ import {OnDestroy} from '@angular/core';
 import {Store} from '@ngrx/store';
 import {ApplicationState} from '../app/state/app.state';
 import {SharedQuery} from '../app/state/shared.state';
-import {Observable} from 'rxjs/index';
+import {Observable} from 'rxjs';
+import {throwError} from 'rxjs/internal/observable/throwError';
 
 /**
  * Created by loicsalou on 16.06.17.
@@ -34,7 +35,7 @@ export abstract class AbstractPersistenceService implements OnDestroy {
 
   protected handleError(message: string, error: any) {
     this.notificationService.error(message, error);
-    return Observable.throwError(error.json().error || 'Firebase error');
+    return throwError(error.json().error || 'Firebase error');
   }
 
   protected subscribeLogin() {
